Add support for deployment.
There are two modes.
The first one, which can be thought of as "pre-packaging",
copies all files with a "deploy" tag to a directory specified by the
property qbs.deployRoot.
The second mode is active if the property qbs.deployInfoFile is set. It
simply gathers all files to be deployed and writes their source and
target paths into that file.

Note that deployment of build products (applications, libraries)
is currently not supported, because the necessary concepts do
not exist yet in qbs.

An in-depth discussion of the overall concept can be found at
http://lists.qt-project.org/pipermail/qbs/2012-February/000025.html.

Change-Id: I1ad5319cbb25cc0e5ca0baa130514e59245fd33a
Reviewed-by: Joerg Bornemann <joerg.bornemann@nokia.com>