ProductModule dependency propagation fixed
Before a ProductModule inherited all modules the product uses.
That's why wrong includePaths etc. were propagated through
the chain of ProductModules.

Now one needs to explicitely add a Depends item for every
module that's used in a ProductModule. This is much more
consistent with the semantics of modules in products.

Example:
    ProductModule {
        cpp.defines: ["FOO"]
    }
is no longer valid. It must be changed to
    ProductModule {
        Depends { name: "cpp" }
        cpp.defines: ["FOO"]
    }

Change-Id: I9b1a42e6bcf56c4fb17e1a19d558fa82d629ff78
Reviewed-by: Marco Bubke <marco.bubke@nokia.com>