Try to optimize the critical path.
Profiling shows that calls to QDir::cleanPath() account to roughly a
third of the time spent in qbs during a build, so we try to call that
function only if necessary. For a complete build of QtCreator, this
patch reduces the number of these calls from ~14 million to ~12 million,
speeding up the scanning process (where they almost exclusively appear)
by a modest seven per cent.

Change-Id: I5aee13e4e54ad4e242d80dc7da557ccb60a232fe
Reviewed-by: Joerg Bornemann <joerg.bornemann@nokia.com>