don't clutter the global object when evaluating probes
We're now setting a special global object when evaluating
probes that contains the imports for TextFile, Process etc.

Unfortunately I did not find a better solution.
For some reason does the prototype chain not work for the activation
object in QtScript.
Also, pushing the imports on to the evaluation context stack breaks when
our script classes come into play.

Change-Id: I31c20206e2526602819d241b81aca2dc595c74ae
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>