do not evaluate all module properties for the project configuration
The project configuration is used to identify build graphs by the
options given on the command line and the current profile setup.
For that we need the properties profile, buildVariant and platform.
All these are given in userconfig passed to the Loader, which is
the current buildconfig. Now we're just using this as the project's
configuration.

Change-Id: I10babcf69105553ee5c4b2497ecb77a81e03d06f
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>