remove multiple QbsEngine support from BuildGraph
Once upon a time there was a developer who was convinced that all parts
of qbs must be run in QtConcurrent context. The consequence was that
every part of qbs had to be prepared to suddenly be run in a different
thread. Especially Angie, the little script engine, was very surprised
when being forced into another thread. It cried rivers and was a woeful
thing to look upon. But then the developer came to the conclusion that
Angie might be comforted by some friends. And so every thread got its
own script engine and Angie was happy to stay in her thread.

Even though this story has a happy end, I find it quite stupid and we
can avoid it.

Change-Id: I93024f8d8d8c025b5301a5547aea4c17214a8e20
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>