Introduce logger class for timing activities.
Instantiating an object of this class before some
interesting activity will log start and end of that
activity as well as how long it took.
The class is designed so that its objects only acquire
resources if they are actually supposed to do something,
as to not introduce unacceptable overhead at low
log levels.

Change-Id: I98b5bf202b5648d96cb12f0e7d8f3fae24029bbd
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>