Move the PersistentObject class into its own header.
At the moment, every declaration of a class inheriting from
PersistentObject has to include the template-heavy persistence.h file,
causing it to transitively be included in pretty much every source file.
By decoupling PersistentObject from that header, we get rid of a bunch
of unneeded compile dependencies, speeding up compilation in general and
especially if something in persistence.h changes.

Change-Id: I1a95390015f75bdfca0d139144fa5c54c7971dc4
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>