remove the "Artifact within a module" feature
By placing an artifact within a module one could add this artifact to
the list of source files of the product that uses this module.
This solution is the half-baked attempt to provide source modules
(see QBS-59). We rather remove the feature for now and reimplement it
properly some time later.

Change-Id: Ia389920843762d108bb6d1c184d93b0986ea315f
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>