fix evaluation of module values from imported JS files
When importing a JS file, a JS object X is created that contains the
evaluated code. Due to the importer's implementation, X is a former
activation object that's been pulled out of a new context.
  When evaluating a Scope property, we call setActivationObject. This
replaces - and thus destroys - the first script value in the scope chain
that is of the internal type "activation object".
  That's why X got destroyed when evaluating qbs module values from code
that lives in X.

We're solving this issue by copying all properties of the activation
object into a fresh and clean JS object. Thus we're avoiding any special
handling of activation objects.

Change-Id: Ibbf3ecca961a781d4a3e0886308f90ae48ecadba
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>