Do product filtering at an earlier stage.
It is unintuitive and confusing to give the Sourceproject
and the Executor a list of projects as part of their build()
operations, when in reality just some products should be built.
So take the "selectedProducts" out of the Buildoptions and
do the filtering outside of the library. This way, we can tell
the API "build these projects" or "build these products" and
it will mean exactly that.

Change-Id: I88860a9cf7f4058145964b1b56b01e7970c9e8d0
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>