Introduce the first version of a public API.
We offer the following services:
    - Build a number of projects.
    - Build a number of products.
    - Retrieve the project structure.
    - Find the executable for a product.
    - Set up an environment to run a shell or
      a target (mainly interesting for our command-line client).

In order not to leak internal data, some parallel data structures
have been introduced for external access to project information
such as source files etc. These are intentionally minimal and
can be extended once use cases arrive.
Note that the Group class has been renamed to ResolvedGroup,
so we can use the simple name for the external API.

Change-Id: I6dc5d7914ee9b232098a55f9d3b0e725debe3bc8
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>