introduce property RuleArtifact.alwaysUpdated
RuleArtifact.alwaysUpdated is a bool property that specifies, if an
output artifact will always be updated by its transformer. The default
value is true.
When finishing the build of artifact A, qbs will retrieve the timestamp
of A, if and only if A.alwaysUpdated is false. Otherwise the actual
timestamp of A is ignored and the current time is used.

When determining whether to run a transformer, only the timestamps of
artifacts that have alwaysUpdated set to true are taken into account.
Transformers whose outputs have all alwaysUpdated == false are always
executed.

Change-Id: I19d2626f17640a813f862980fafd507a13fcca95
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>