Get rid of some raw pointer members ...
... where they actually point to objects held in
shared pointers elsewhere. In such cases, use
shared pointers when possible and weak pointers
when necessary. Beside better self-documentatiom,
this also enables us to get rid of some ugly loops
that were needed to get from the raw pointer value
back to the shared equivalent.
(We ignore members held in the Artifact class for now
as to not complicate the patch.)

Change-Id: Icc47c790835c4d144ddab5c1de65fc2c5de5e4c7
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>