Move build root information out of BuildGraph.
With the current implementation, we would need to create a new QbsEngine
object for every project file.
Instead, attach a build directory to every ResolvedProject. Also remove
ResolvedProduct::buildDirectory, since it is the same for all products
in a project.
Along the way:
    - Integrate BuildProject::restoreBuildGraph() into
BuildProject::loadProject(). The latter currently just forwards to the
former for no apparent reason.
    - Remove some checks for things that can never happen.
    - Remove unused metatype declaration for internal class.

Change-Id: Ib97b37ddfdc0f4f4334449f64cde1a33eda4316d
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>