Make the public API asynchronous.
This simplifies things a lot for IDEs, which are no longer
forced to run a dedicated thread for a QbsEngine, as
everything on the calling side can now happen in the main loop.
Getting rid of blocking operations also allows for more flexibility
in that it is now possible to load one project while
another is building etc.
This patch also moves all the high-level types into src/lib/api/.
Until now, they were scattered around language/ and tools/, where
they do not belong.
Very few things had to be touched in the back-end; most notable
is the fact that BuildGraph's ScriptEngine is now settable,
because we must exchange it between resolving and building,
as the former operation runs in its own thread.

Change-Id: I5c8a7cc3517ebc2489678c0d5b1068972997e930
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>