Sanitize the Executor code flow.
The current behavior is not exactly well-defined, mostly
due to signals calling back on the Executor while it is
calling a function on the object emitting the signal (e.g.
waitForFinished()). These traps are now gone.
Also make sure error messages are in a sensible order
when a build is canceled.
In addition, remove unused functionality.

Change-Id: Icf35eb378995a55b5ba4fc6caea9492eccb43527
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>