Fix renaming files to be installed that are matched by a wildcard.
The problem is that files matched by an "install" rule are currently not
added to the target artifacts if a build graph already exists and they
are "implicitly added" by adding or renaming a file in a group matched
by a wildcard (i.e. the change happens without touching the project
file).

Task-number: QBS-146
Change-Id: Iec499de8dd88e3ac599b7435ec441aad7689f55a
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>