Fix semantics of recursive wildcards.
Consider this:
    Group {
        files: "x/*"
        qbs.installDir: "x"
        recursive: true
        fileTags: "install"
    }
Now let's assume there is a file x/y/z.txt. The current implementation
expands the whole path and copies the resulting file to the base install
dir, so that z.txt ends up in <install root>/x/z.txt. The new
implementation does not descend into y, so that the structure in the
installation directory is x/y/z.txt, as one would expect.

Change-Id: I31062b743e5b4d6938bbc58090f36bb02d2bd094
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>