Re-do the wildcard implementation.
This feature is in desperate need of simplification, since the
current semantics are not clearly defined. The main
culprit for that is the ability to use wildcards in
directory names, which leads to all sorts of
problems, especially in combination with the "recursive" flag.
We therefore disallow that now: Wildcards can only be used
to match actual files. This results in clear semantics that
are easily understood and can actually be documented.

Change-Id: If8a950e1c1a362ce9fb57a1c9789a83ea7c8b3ba
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>