introduce property Group.fileTagsFilter
With fileTagsFilter one can set properties on a certain set of
generated artifacts. This property is a list of file tags.
All artifacts that have file tags that are in that list will get
the properties that are bound in the rest of the item.

Example:

    Group {
        fileTagsFilter: ["dynamiclibrary"]
        qbs.installDir: "mylibs"
    }

Group.files and Group.fileTagsFilter are exclusive.

Change-Id: I764e0a823f2cda4efe4ac3c985ecde5e178bab7b
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>