Check whether stored build graph is really for our project.
At the moment, we accept build graphs for any project and claim "Build
done" if it's newer than the project file.
With this patch, we error out in such a case by default. We do not
simply discard the existing build, because presumably the user made a
mistake and does not want the existing build to be overwritten by one
for a completely different project.
If a user is sure it's the same project after all (e.g. the sources have
moved or the project file was renamed), the "--force" option can be
used.
If a user really wants to throw away the build for the other project,
that must be done manually.

Change-Id: I1194fcc4695632a6b96789bc8032f9d3b3a55ce3
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>