do not scan qrc dependencies recursively
The files referenced in a qrc file must not be scanned for further
dependencies.

Assume the following situation:
    - foo.qrc references bar.cpp,
    - bar.cpp includes bar.h
We do not want to regenerate qrc_foo.cpp when bar.h changed.

ScannerPlugin::usesCppIncludePaths has been turned into the flag
ScannerUsesCppIncludePaths. We introduce the scanner flag
ScannerRecursiveDependencies. If this flag is set, then every scan
result is added to the list of file to scan.
Typically, this will be set for include-file-like dependencies.

Change-Id: I99e04f554fc34db3e0b90610bb3d20f7e6000c57
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>