Place the executable in the application bundle directly.
As a side effect, this fixes installation of application bundles
(the entire .app directory is copied, not just the executable).

The prepare command remains in the applicationbundle Rule to let
the user know that an application bundle has been generated,
though it performs no actions other than printing that message.

Task-number: QBS-186
Change-Id: I61c5d5c3fb6e64251026a346a688563546426dbb
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>