cache Rule.prepare scripts as function objects
For Rule.prepare a function object is created on demand. This function
object is called on every rule application. The QScriptProgram cache
turns out to be quite useless.

This caching approach saves roughly 5% of time of the whole resolve
phase.

Magic persistence number changed, because the createCommands
method now expects prepare scripts in function form.

Change-Id: Ie531dcfb5f21578548a5961ddaa6086ae441a2b0
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>