Add XIB compiler.
XIB files are now compiled into NIBs and placed in the bundle.

A manual test has been added for testing this functionality. Note that
the project was generated using one of the Xcode templates; this test
will therefore be useful for ensuring that QBS can generate an
identical application bundle to Xcode in general.

Task-number: QBS-202
Change-Id: Ie098327a65c5340dd554cb7d6f49415efd1def27
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>