Fix handling of minimum Windows version and subsystem linker flags.
Both MSVC and MinGW now only pass their respective subsystem flags if
product.consoleApplication is not undefined - the linkers use console by
default.

MSVC: as an exception to the above, a subsystem flag must be passed
regardless of the value of product.consoleApplication if a minimum
system version was specified - in this case subsystem will be set to
console.

MinGW now correctly handles setting the minimum subsystem/OS version; it
was previously using totally nonexistent flags due to poor assumptions
about its MSVC compatibility.

The test case which tests this functionality was also broken, and is now
fixed as well.

Task-number: QBS-244
Change-Id: Ibca29ce673a81f1231d364d5b6e6875a462b379c
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>