fix memory leak in loader
Cycles in the graph of item objects prevent us from using shared
pointers. The whole structure is too complex to guarantee the absence
of cycles in every case.
To fix that we remove the smart pointers and introduce the ItemPool
as the sole owner of all items of a project.

Task-number: QBS-269

Change-Id: I53c55216598086b82d00ede08c56395062384834
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>