Centralize Info.plist/bundle handling code.
The differences in plist/bundle handling between OS X and iOS are
extremely minimal, so there is no reason to duplicate such a large block
of code between their respective module files.

As a side effect, this implements support for dictionaries and arrays
in OS X plists as well as the defaults handling that already existed
for iOS.

As an additional side effect, QBS-272 is no longer an issue.

Change-Id: I39be5681f4d4fba7369d55da720ff8130ff52cba
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>