Introduce a series of functions for working with Darwin bundles.
These functions abstract away the physical structure of the bundles,
allowing the user to query for the paths to particular files and
subdirectories within the bundle in a manner that is independent of
the bundle type and platform.

The fileName properties of relevant Artifacts have been updated to take
advantage of the new BundleTools functions.

Change-Id: I051d623ac7eae9cacf90495c662ad4ad1cf6128f
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>