split overridden values off the build configuration
Property values in profiles now serve as default values for module
values. Property values given on the command line override properties
that are specified in files.
This way it is possible to specify a property value in a profile but
override it in a file or on the command line.
The hierarchy is profile < file < command line.

Change-Id: I786ef5a215312779ed1b79ed64f856b82c612403
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>