assign the product's build config to rule outputs
Before: Every output artifact of a rule output inherits the build config
of an arbitrary input artifact if the rule is not a multiplex rule.
Now: Every output artifact of a rule gets the product's build config.

The former behavior falls apart if a rule has multiple inputs and
creates inconsistency between multiplex and non-multiplex rules.
It also has the effect that object files get installed if source files
are marked with "install: true".

Change-Id: I59de3d037b20e9d3c21c244da0d3547410e9e4da
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>