Add shortcuts to some equality checks.
Comparing the pointers first is both faster in case the two references
actually point to the same object and also allows us to do a null
pointer check before accessing any members.
Note that doing the null pointer check in operator== means that we have
already applied the operator* to the object. which is a bit... dubious.
However, since no actual access to any data happens, this is very
unlikely to cause problems, and it has the advantage that we will catch
all comparisons, not just the ones that we would explicitly route
through some wrapper function we would write otherwise.

Change-Id: Ia63508ae925520565956086403fa1eb6494e3de8
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>