Improve the way that the WiX installation is found.
Now that we have registry access, we search the registry for the key
corresponding to the latest known WiX version, then determine the WiX
installation path and version number from that. The latter is now
available as a set of properties.

This searching code is also now not broken for 32-bit Windows...

Task-number: QBS-413
Change-Id: I2fae8164b4aa073c1b0b980f110af38a1c4ddc66
Reviewed-by: Jake Petroules <jake.petroules@petroules.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>