Allow lists of patterns in the FileTagger item.
Rationale: It is not uncommon that the same tag is associated
with more than one file pattern; more common, in fact, than the
other way around. It therefore seems silly to force module authors
to create one FileTagger item per file extension.
For semantic consistency, the "pattern" property gets renamed to
"patterns". The old name is still supported in this minor version.

Change-Id: Ia7f744b0903462517582fcb549e2dab004c81dd5
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>