Fix file tagger pattern matching.
When first creating a FileTagger object, we implicitly set the pattern
syntax to "RegExp". On restoring, we set it to "WildCard".
We now use wildcards always, as the documentation says.
Note that the only reason the bug has not manifested itself yet is that
the shortcut in FileInfo::globMatches() catches all our existing file
tags.

Change-Id: Ibedb54fb3a890ae6ab35a8f073d85cc4ebfc36bc
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>