Fix look-up of scalar module properties.
These must not be searched for recursively. For instance, if such a
property is set to different values in different instantiating modules,
there is no sensible heuristic which value should be used. The current
implementation just takes the one found first, which is just as wrong as
any other solution that looks beyond what is set directly in the product
or artifact.

Change-Id: I65c29538a437872e2d1e1c4043f2c5cfa301ebb9
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>