Doc: Use the \QBS macro for the product name
This ensures that the name is always written in the
same way and makes it easy to change, as only the value
of the macro needs to be changed.

Removed the product name from most titles as macros cannot
be used in titles. It would break linking to titles.

Please always use the macro in the future.

Change-Id: I116abffc316edbb861034f5e431ca8e27b5a5d03
Reviewed-by: Jake Petroules <jake.petroules@petroules.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>