QBS: Add a new linkerScripts property for the CppModule
Linkers of the majority Embedded projects (like the "bare metal" without
an OS) require of the script files which contain directives for placement
a binary code of FW to the desired address space of the target MCU.

When using a QBS project it can be implemented by adding to the
"cpp.linkerFlags" property, e.g. :

cpp.linkerFlags: [ "-T", "<path/to/script/file>" ]

But this solution has a shortcomings related to need to specify a full
paths to the scripts files and so on.

Therefore was introduced a new "cpp.linkerScripts" property which requires
specifying of the list of used scripts files with the relative paths
concerning of the *.qbs project file, e.g. :

cpp.linkerScripts: [ "<scriptfile1>", "<scriptfile2>" ]

Note: This property is implemented only for the GCC compiler module.

Change-Id: I9f30a054e615a1cef8ca022352706eb3448bc7cd
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>