fix OS X dsym file tags
The usage of the same dsym file tag in the app bundle and framework
bundle rules led to two separate rule paths.
One was creating a dylib
    cpp -> obj -> dylib -> dsym -> framework
the other created an app
    cpp -> obj -> app -> dsym -> app bundle
Using distinct file tags for application and dylib dsyms fixes this
conflict. We should detect and handle situations like this in a
subsequent commit.

Task-number: QBS-384
Change-Id: I931b6a4be929468e4abdc2385aa38d8c224b32c6
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>