Introduce tool for fuzzy testing.
Incremental builds are the likeliest source of problems, as well as the
one that is the hardest to reproduce. This new tool turns a given project
into a quasi-autotest by checking out random commits and doing incremental
builds on them. In case of a build error that is not reproducible with
a clean build, it aborts and reports the offending sequence of commits.
Any qbs project that comes in a git repository is supported.

Change-Id: I9f6f9172ce03140341ed25859092a8ed32d4a17d
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>