Do not share product properties with artifacts.
Since products potentially share their property maps with their groups,
artifacts get product properties such as "name" and
"destinationDirectory". This is strange and can mess with change
tracking. Therefore we split up the product's property map in two: One
contains the module properties and is potentially shared with groups,
the other has the actual product properties and is not shared.
This patch also removes a workaround from an autotest that was required
until now to prevent a false positive in change tracking.

Change-Id: Ia1f1f0ce32669fd893a99f809753df526bf1442a
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>