remove JSSourceValue::Alternative::conditionScopeItem
This member held the scope item for the condition of JSSourceValue
alternatives (Properties items). However, the condition scope is
always the same as the scope item of the JSSourceValue itself.
The missing part was that we need to make sure that the scopes always
have a valid FileContext pointer.

Removing this member made it possible to remove
ItemReaderResult::conditionalValuesPerScopeItem as well.

This patch will enable us to implement item inheritance by using the
prototype chain.

Change-Id: I77e976cdfc5af072e35d430b055b6ae1fac930a4
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>