fix FileInfo::isFileCaseCorrect on OS X
This function is supposed to compare file *names* and not paths.
If it gets passed a path to a soft link like /tmp/foo, then
the canonical file path of that will be /private/tmp/foo.
We must not compare the whole path but only its file name
portion.

Task-number: QTCREATORBUG-11875
Change-Id: I9724c6e45057aa9a80c3a9bd00c5fc225799b6c5
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>