Fix potential assertion in Executor.
In the CommandExecutor and ExecutorJob classes, we follow QProcess'
glorious example and report error() and finished() separately. In the
case of a process crashing, the current code can lead to a slot in the
Executor being called twice, the second time in an unexpected state.
This patch fixes that particular bug.

Change-Id: I2f75a5426bc8e14912d0c796fea5c905afd0e628
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>