cpp module: Make library symlink artifact creation dynamic.
Currently, the build graph artifacts for symbolic links to a dynamic
library (such as libfoo.so.1.0) are always created, even if the library
does not have a version number. They get then merged with the actual
library artifact due to the identical file path, which is weird.
Instead, create such an artifact only if the corresponding symbolic link
will really be created by the command.

Change-Id: I6af903132560b1ea6ebd4f0d073a37cf845e7533
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>